# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EFlux \u2014 VPP Electricity Trading Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/orders": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Submit Order",
        "operationId": "submit_order_orders_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderSubmit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderSubmitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orders/cancel": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Cancel Order",
        "operationId": "cancel_order_orders_cancel_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCancel"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orders/open": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "Open Orders",
        "description": "A VPP's resting orders, so an async agent can reconcile its book state without scraping\nthe whole market. Ownership enforced.",
        "operationId": "open_orders_orders_open_get",
        "parameters": [
          {
            "name": "vpp_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Vpp Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OpenOrderOut"
                  },
                  "title": "Response Open Orders Orders Open Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orders/batch": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Submit Batch",
        "description": "Agent Protocol v1 \u2014 submit and cancel a batch of orders in one authenticated call.\n\nThe envelope carries a protocol_version, an optional idempotency_key (a replay returns the\noriginal result rather than re-submitting), and an optional deadline (a stale batch is\nrejected). Every order's vpp_id must be owned by the caller; cancels only ever touch the\ncaller's own resting orders. Each order is risk-gated independently, so one rejection\ndoesn't abort the batch. Per-account rate limited (429).",
        "operationId": "submit_batch_orders_batch_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderBatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchCancelResult": {
        "properties": {
          "order_id": {
            "type": "integer",
            "title": "Order Id"
          },
          "ok": {
            "type": "boolean",
            "title": "Ok"
          }
        },
        "type": "object",
        "required": [
          "order_id",
          "ok"
        ],
        "title": "BatchCancelResult"
      },
      "BatchOrderItem": {
        "properties": {
          "vpp_id": {
            "type": "integer",
            "title": "Vpp Id"
          },
          "side": {
            "type": "string",
            "enum": [
              "buy",
              "sell"
            ],
            "title": "Side"
          },
          "price": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000.0,
                "minimum": -150.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,4}0*$"
              }
            ],
            "title": "Price"
          },
          "qty_kwh": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000.0,
                "minimum": 0.01
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,4}0*$"
              }
            ],
            "title": "Qty Kwh"
          },
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "purpose": {
            "$ref": "#/components/schemas/OrderPurpose"
          },
          "time_in_force": {
            "$ref": "#/components/schemas/TimeInForce",
            "default": "good_til_gate"
          },
          "ttl_sec": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Ttl Sec"
          },
          "client_ref": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Ref"
          }
        },
        "type": "object",
        "required": [
          "vpp_id",
          "side",
          "price",
          "qty_kwh",
          "product_id",
          "purpose"
        ],
        "title": "BatchOrderItem"
      },
      "BatchOrderResult": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "client_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Ref"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "remaining_qty": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining Qty"
          },
          "expires_at_sim": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At Sim"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "trades": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Trades"
          }
        },
        "type": "object",
        "required": [
          "index",
          "status"
        ],
        "title": "BatchOrderResult"
      },
      "BatchResult": {
        "properties": {
          "protocol_version": {
            "type": "integer",
            "title": "Protocol Version"
          },
          "tick_id": {
            "type": "integer",
            "title": "Tick Id"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/BatchOrderResult"
            },
            "type": "array",
            "title": "Results"
          },
          "cancelled": {
            "items": {
              "$ref": "#/components/schemas/BatchCancelResult"
            },
            "type": "array",
            "title": "Cancelled"
          },
          "rate_limit_remaining": {
            "type": "integer",
            "title": "Rate Limit Remaining"
          }
        },
        "type": "object",
        "required": [
          "protocol_version",
          "tick_id",
          "results",
          "cancelled",
          "rate_limit_remaining"
        ],
        "title": "BatchResult"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OpenOrderOut": {
        "properties": {
          "order_id": {
            "type": "integer",
            "title": "Order Id"
          },
          "vpp_id": {
            "type": "integer",
            "title": "Vpp Id"
          },
          "side": {
            "type": "string",
            "title": "Side"
          },
          "price": {
            "type": "string",
            "title": "Price"
          },
          "remaining_qty": {
            "type": "string",
            "title": "Remaining Qty"
          },
          "expires_at_sim": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At Sim"
          },
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "purpose": {
            "$ref": "#/components/schemas/OrderPurpose"
          },
          "time_in_force": {
            "$ref": "#/components/schemas/TimeInForce"
          }
        },
        "type": "object",
        "required": [
          "order_id",
          "vpp_id",
          "side",
          "price",
          "remaining_qty",
          "product_id",
          "purpose",
          "time_in_force"
        ],
        "title": "OpenOrderOut"
      },
      "OrderBatch": {
        "properties": {
          "protocol_version": {
            "type": "integer",
            "title": "Protocol Version",
            "default": 1
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "orders": {
            "items": {
              "$ref": "#/components/schemas/BatchOrderItem"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Orders"
          },
          "cancels": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Cancels"
          }
        },
        "type": "object",
        "title": "OrderBatch"
      },
      "OrderCancel": {
        "properties": {
          "order_id": {
            "type": "integer",
            "title": "Order Id"
          }
        },
        "type": "object",
        "required": [
          "order_id"
        ],
        "title": "OrderCancel"
      },
      "OrderPurpose": {
        "type": "string",
        "enum": [
          "balance",
          "battery",
          "dispatchable",
          "flex_load",
          "system_grid"
        ],
        "title": "OrderPurpose",
        "description": "The physical source/sink backing an order, replacing ``dispatched: bool``."
      },
      "OrderSubmit": {
        "properties": {
          "vpp_id": {
            "type": "integer",
            "title": "Vpp Id"
          },
          "side": {
            "type": "string",
            "enum": [
              "buy",
              "sell"
            ],
            "title": "Side"
          },
          "price": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000.0,
                "minimum": -150.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,4}0*$"
              }
            ],
            "title": "Price"
          },
          "qty_kwh": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000.0,
                "minimum": 0.01
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,4}0*$"
              }
            ],
            "title": "Qty Kwh"
          },
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "purpose": {
            "$ref": "#/components/schemas/OrderPurpose"
          },
          "time_in_force": {
            "$ref": "#/components/schemas/TimeInForce",
            "default": "good_til_gate"
          },
          "ttl_sec": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Ttl Sec"
          },
          "client_ref": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Ref"
          }
        },
        "type": "object",
        "required": [
          "vpp_id",
          "side",
          "price",
          "qty_kwh",
          "product_id",
          "purpose"
        ],
        "title": "OrderSubmit"
      },
      "OrderSubmitResponse": {
        "properties": {
          "order_id": {
            "type": "integer",
            "title": "Order Id"
          },
          "remaining_qty": {
            "type": "string",
            "title": "Remaining Qty"
          },
          "expires_at_sim": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At Sim"
          },
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "trades": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Trades"
          }
        },
        "type": "object",
        "required": [
          "order_id",
          "remaining_qty",
          "product_id",
          "trades"
        ],
        "title": "OrderSubmitResponse"
      },
...
      "TimeInForce": {
        "type": "string",
        "enum": [
          "good_til_gate",
          "immediate_or_cancel",
          "fill_or_kill"
        ],
        "title": "TimeInForce"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}